resmon: add support for notification file
We have support to watch the kubelet state directory
for state file change, triggering podresources poll.
This is what we call "smart polling".

In some cases, it may be unpractical to allow RTE to
access the kubelet state directory, even in read only.

To still enable the smart polling, we add support for
a notification files, to be created in a custom-provided
directory. We intentionally keep the mechanism as simple
as possible for clients: just touching this file
(see man 1 touch) triggers a podresources poll.

Signed-off-by: Francesco Romani <fromani@redhat.com>